create spawn module
