--- thetawave-mobs/src/plugin.rs
@@ -1,28 +1,0 @@
-use bevy::{
-    app::{Plugin, Update},
-    ecs::schedule::{Condition, IntoScheduleConfigs, common_conditions::on_event},
-    state::condition::in_state,
-};
-use thetawave_states::{AppState, GameState};
-
-use crate::{
-    SpawnMobEvent, attributes::ThetawaveAttributesPlugin, behavior::ThetawaveMobBehaviorPlugin,
-    systems::spawn_mob_system,
-};
-
-pub struct ThetawaveMobsPlugin;
-
-impl Plugin for ThetawaveMobsPlugin {
-    fn build(&self, app: &mut bevy::prelude::App) {
-        app.add_plugins((ThetawaveMobBehaviorPlugin, ThetawaveAttributesPlugin));
-
-        app.add_systems(
-            Update,
-            spawn_mob_system.run_if(
-                in_state(AppState::Game)
-                    .and(in_state(GameState::Playing).and(on_event::<SpawnMobEvent>)),
-            ),
-        )
-        .add_event::<SpawnMobEvent>();
-    }
-}
